core: add NMDevIdxIndex as a storage for information about NMDevices
Currently, devices are tracked by NMManager with a GSList.

Introduce NMDevIdxIndex as a container class to keep track of NMDevices.
Eventually it can be used by NMManager to replace the existing data
structures with the indexed one.

Also, NMPolicy needs to track information for each NMDevice-NMSettingsConnection
pair. NMDevIdxIndex will be used for that too.

The point is to have a class that has mostly O(1) access time for most
operations. For that NMDevIdxIndex is a mixture of hash tables and links
entries together with a CList.

Also, by moving the indexing code to a separate class, we can make
it testable independent of NMManager or NMPolicy.